# source/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document test API Locofull",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "tien.nguyen@saver.jp"
        },
        "license": {
            "name": "Saver",
            "url": "https://www.saver.jp/"
        },
        "version": "1.0.0"
    },
    "basePath": "/backend/api",
    "paths": {
...
        "/connector/request": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Connector request for tranfer money",
                "parameters": [
                    {
                        "name": "introduction_status_id",
                        "in": "formData",
                        "description": "Introduction Status Id",
                        "required": true,
                        "type": "number",
                        "default": "1"
                    },
                    {
                        "name": "api_token",
                        "in": "formData",
                        "description": "API token",
                        "required": true,
                        "type": "string",
                        "default": "U1n8N6yLQbDNhQL0Gl3y1WXHV0fp9XOudgpGcYQy14PGVfV0LMemtMegto6m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{<pre>&emsp;  'success': true | false,\r\n                                                        &emsp;  'data': null,\r\n                                                        &emsp;  'error': null | ...</pre>}\r\n                                                        "
                    },
                    "404": {
                        "description": "404 page"
                    }
                }
            }
        },
...
    },
...
}